+++ src/aui-form-builder/tests/unit/js/aui-form-builder-field-base-tests.js
@@ -97,6 +97,17 @@ YUI.add('aui-form-builder-field-base-tests', function(Y) {
             Y.Assert.isTrue(container.one('.form-builder-field-toolbar').hasClass('hide'));
         },
 
+        'should check if toolbar is visible': function() {
+            this._field = new TestField();
+            Y.one('#container').append(this._field.get('node'));
+
+            this._field.toggleToolbar(true);
+            Y.Assert.isTrue(this._field.isToolbarVisible());
+
+            this._field.toggleToolbar(false);
+            Y.Assert.isFalse(this._field.isToolbarVisible());
+        },
+
         'should toggle the configuration button': function() {
             var container = Y.Node.create('<div></div>');
 
@@ -171,6 +182,35 @@ YUI.add('aui-form-builder-field-base-tests', function(Y) {
 
             Y.Assert.isNotNull(this._field.get('content'));
             Y.Assert.isTrue(this._field.get('content').hasClass('form-builder-field'));
+        },
+
+        'should only allow array of fields as the nestedFields attribute': function() {
+            this._field = new TestField();
+
+            Y.Assert.areEqual(0, this._field.get('nestedFields').length);
+
+            this._field.set('nestedFields', 1);
+            Y.Assert.areEqual(0, this._field.get('nestedFields').length);
+
+            this._field.set('nestedFields', [1, 2]);
+            Y.Assert.areEqual(0, this._field.get('nestedFields').length);
+
+            this._field.set('nestedFields', [1, new TestField()]);
+            Y.Assert.areEqual(0, this._field.get('nestedFields').length);
+
+            this._field.set('nestedFields', [new TestField(), new TestField()]);
+            Y.Assert.areEqual(2, this._field.get('nestedFields').length);
+        },
+
+        'should render nested fields inside content': function() {
+            this._field = new TestField({
+                nestedFields: [new TestField(), new TestField()]
+            });
+
+            Y.Assert.areEqual(2, this._field.get('content').all('.form-builder-field').size());
+
+            this._field.set('nestedFields', [new TestField()]);
+            Y.Assert.areEqual(1, this._field.get('content').all('.form-builder-field').size());
         }
     }));